Implement DocsGenerate and DocsGenReferences tests
