+++ tests/functional/adapter/test_basic.py
@@ -14,8 +14,9 @@ from dbt.tests.adapter.basic.test_generic_tests import BaseGenericTests
 from dbt.tests.adapter.basic.test_snapshot_check_cols import BaseSnapshotCheckCols
 from dbt.tests.adapter.basic.test_snapshot_timestamp import BaseSnapshotTimestamp
 from dbt.tests.adapter.basic.test_adapter_methods import BaseAdapterMethod
-from dbt.tests.adapter.basic.test_docs_generate import BaseDocsGenerate
-from dbt.tests.adapter.basic.expected_catalog import base_expected_catalog, no_stats
+from dbt.tests.adapter.basic.test_docs_generate import BaseDocsGenerate, BaseDocsGenReferences
+from dbt.tests.adapter.basic.expected_catalog import base_expected_catalog, no_stats, expected_references_catalog
+from tests.functional.adapter.expected_stats import redshift_stats
 
 from dbt.tests.adapter.basic.files import seeds_base_csv, seeds_added_csv, seeds_newcolumns_csv
 
@@ -87,72 +88,6 @@ class TestBaseAdapterMethod(BaseAdapterMethod):
     pass
 
 
-def redshift_stats():
-    return {
-        "has_stats": {
-            "id": "has_stats",
-            "label": "Has Stats?",
-            "value": True,
-            "description": "Indicates whether there are statistics for this table",
-            "include": False
-        },
-        "encoded": {
-            "id": "encoded",
-            "label": "Encoded",
-            "value": AnyStringWith('Y'),
-            "description": "Indicates whether any column in the table has compression encoding defined.",
-            "include": True
-        },
-        "diststyle": {
-            "id": "diststyle",
-            "label": "Dist Style",
-            "value": AnyStringWith('AUTO'),
-            "description": "Distribution style or distribution key column, if key distribution is defined.",
-            "include": True
-        },
-        "max_varchar": {
-            "id": "max_varchar",
-            "label": "Max Varchar",
-            "value": AnyFloat(),
-            "description": "Size of the largest column that uses a VARCHAR data type.",
-            "include": True
-        },
-        "size": {
-            "id": "size",
-            "label": "Approximate Size",
-            "value": AnyFloat(),
-            "description": "Approximate size of the table, calculated from a count of 1MB blocks",
-            "include": True
-        },
-        'sortkey1': {
-            'id': 'sortkey1',
-            'label': 'Sort Key 1',
-            'value': AnyStringWith('AUTO'),
-            'description': 'First column in the sort key.',
-            'include': True,
-        },
-        "pct_used": {
-            "id": "pct_used",
-            "label": "Disk Utilization",
-            "value": AnyFloat(),
-            "description": "Percent of available space that is used by the table.",
-            "include": True
-        },
-        "stats_off": {
-            "id": "stats_off",
-            "label": "Stats Off",
-            "value": AnyFloat(),
-            "description": "Number that indicates how stale the table statistics are; 0 is current, 100 is out of date.",
-            "include": True
-        },
-        "rows": {
-            "id": "rows",
-            "label": "Approximate Row Count",
-            "value": AnyFloat(),
-            "description": "Approximate number of rows in the table. This value includes rows marked for deletion, but not yet vacuumed.",
-            "include": True
-        },
-    }
 
 class TestDocsGenerateRedshift(BaseDocsGenerate):
     @pytest.fixture(scope="class")                               
@@ -161,11 +96,29 @@ class TestDocsGenerateRedshift(BaseDocsGenerate):
             project,                               
             role=profile_user,                    
             id_type="integer",                      
-            text_type=AnyStringWith('character varying'),
+            text_type=AnyStringWith("character varying"),
             time_type="timestamp without time zone",
             view_type="VIEW",                
             table_type="BASE TABLE",                 
             model_stats=no_stats(),
             seed_stats=redshift_stats(),
         )             
+
+
+class TestDocsGenReferencesRedshift(BaseDocsGenReferences):
+    @pytest.fixture(scope="class")
+    def expected_catalog(self, project, profile_user):
+        return expected_references_catalog(
+            project,
+            role=profile_user,
+            id_type="integer",
+            text_type=AnyStringWith("character varying"),
+            time_type="timestamp without time zone",
+            bigint_type="bigint",
+            view_type="VIEW",
+            table_type="BASE TABLE",
+            model_stats=redshift_stats(),
+            seed_stats=redshift_stats(),
+            view_summary_stats=no_stats(),
+        )
